chore(api-redflags): Fix failing tests
