+++ .gitignore
@@ -2,3 +2,5 @@ venv
 .vscode
 __pycache__
 *.pyc
+htmlcov
+.coverage
