+++ api/auth/routes.py
@@ -11,8 +11,3 @@ def signup():
 @api.route("/login", methods=['POST'])
 def login():
     return user_controller.user_login()
-
-@api.route("/users", methods=['GET'])
-@jwt_required
-def get_users():
-    return user_controller.fetch_all_users()
